requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
grpcio==1.62.0
grpcio-tools==1.62.0
idna==3.6
praw==7.7.1
prawcore==2.4.0
protobuf==4.25.3
python-dotenv==1.0.1
pywatchman==1.4.1
requests==2.31.0
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.7.0
